backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
opencv-python-headless==4.12.0.88
numpy>=2.0,<3
pytesseract==0.3.10
easyocr==1.7.1
python-multipart==0.0.6
google-generativeai==0.3.2
torch==2.8.0
torchvision==0.23.0